Refactoring: GeoDataRiverSurvey::EditCrosssectionCommand is moved to separated files.
