+++ libs/geodata/riversurvey/geodatariversurveycrosssectionwindow.h
@@ -2,7 +2,6 @@
 #define GEODATARIVERSURVEYCROSSSECTIONWINDOW_H
 
 #include "gd_riversurvey_global.h"
-#include "geodatariversurvey.h"
 #include "geodatarivercrosssection.h"
 
 #include <guicore/base/snapshotenabledwindowinterface.h>
@@ -10,9 +9,9 @@
 
 #include <QMainWindow>
 #include <QList>
-#include <QUndoCommand>
 
 class GeoDataRiverPathPoint;
+class GeoDataRiverSurvey;
 class GeoDataRiverSurveyCrosssectionWindowGraphicsView;
 class GeoDataRiverSurveyCrosssectionWindowProjectDataItem;
 class PreProcessorGeoDataGroupDataItemInterface;
@@ -141,25 +140,6 @@ private:
 
 public:
 	friend class GeoDataRiverSurveyCrosssectionWindowGraphicsView;
-};
-
-class GeoDataRiverSurvey::EditCrosssectionCommand : public QUndoCommand
-{
-
-public:
-	EditCrosssectionCommand(bool apply, const QString& title, GeoDataRiverPathPoint* p, const GeoDataRiverCrosssection::AltitudeList& after, const GeoDataRiverCrosssection::AltitudeList& before, GeoDataRiverSurveyCrosssectionWindow* w, GeoDataRiverSurvey* rs, bool tableaction = false, QUndoCommand* parentcommand = nullptr);
-	void redo() override;
-	void undo() override;
-
-private:
-	bool m_apply;
-	bool m_first;
-	bool m_tableaction;
-	GeoDataRiverPathPoint* m_point;
-	GeoDataRiverCrosssection::AltitudeList m_before;
-	GeoDataRiverCrosssection::AltitudeList m_after;
-	GeoDataRiverSurveyCrosssectionWindow* m_window;
-	GeoDataRiverSurvey* m_rs;
 };
 
 #if _DEBUG
